First stage, web server starting and connecting to the ODB
