+++ modules/web/server/src/main/resources/app.conf
@@ -26,6 +26,11 @@ smart-gcal {
     smart-gcal-dir = /tmp/smartgcal
 }
 
+lucuma-sso {
+  service-token = "DummyToken"
+  service-token = ${?ODB_SERVICE_JWT}
+}
+
 # Web server related configuration
 web-server {
     # Interface to listen on, 0.0.0.0 listens in all interfaces, production instances should be more restrictive
@@ -41,7 +46,7 @@ web-server {
 # Configuration of the observe engine
 observe-engine {
     # host for the odb
-    odb = "wss://lucuma-odb-development.herokuapp.com/ws"
+    odb = "wss://lucuma-postgres-odb-dev.herokuapp.com/ws"
     dhs-server = "http://cpodhsxx:9090/axis2/services/dhs/images"
     # Tells Observe how to interact with a system:
     #   full: connect and command the system
